+++ TM4C123_GPIO_LCD1602/xDriver_MCU/Driver_Source/SYSTICK/SYSTICK_Driver/SYSTICK_Init.c
@@ -14,10 +14,10 @@ SysTick_nSTATUS SysTick__enInitTick(uint32_t u32Tick, SCB_nSHPR enPriority, SysT
 {
     SysTick_nSTATUS enReturn = SysTick_enERROR;
     uint32_t u32SysTickFrequency=0;
-    float fSysTickFrequency=0,fSysTickUs=1000000.0;
+    float fSysTickFrequency=0u,fSysTickUs=1000000.0;
 
-    SysTick_CSR_R=0;
-    SysTick_CSR->CLKSOURCE=(uint8_t)enClockSource&SysTick_CSR_CLKSOURCE_MASK;
+    SysTick_CSR_R=0u;
+    SysTick_CSR->CLKSOURCE=(uint32_t)enClockSource&SysTick_CSR_CLKSOURCE_MASK;
     if((uint8_t)(enClockSource&SysTick_CSR_CLKSOURCE_MASK) == SysTick_enPIOSC4)
     {
         u32SysTickFrequency=SysTick_PIOSC4;
@@ -27,16 +27,16 @@ SysTick_nSTATUS SysTick__enInitTick(uint32_t u32Tick, SCB_nSHPR enPriority, SysT
         u32SysTickFrequency=SYSCTL__u32GetClock();
     }
 
-    if(u32Tick >40 )
+    if(u32Tick >40u )
     {
-        if(u32Tick>(SysTick_VALUEMAX+1))
+        if(u32Tick>(SysTick_VALUEMAX+1u))
         {
-            u32Tick=(SysTick_VALUEMAX+1);
+            u32Tick=(SysTick_VALUEMAX+1u);
         }
         SysTick__vClearFreeCount();
         SysTick__vClearFreeCountOv();
         SysTick__vSetFreqBase(u32SysTickFrequency);
-        SysTick__vSetTimeUsBase(1000000.0/(float)u32SysTickFrequency);
+        SysTick__vSetTimeUsBase(1000000.0f/(float)u32SysTickFrequency);
 
         fSysTickFrequency=(float)u32SysTickFrequency;
         fSysTickFrequency/=(float)u32Tick;
@@ -47,9 +47,9 @@ SysTick_nSTATUS SysTick__enInitTick(uint32_t u32Tick, SCB_nSHPR enPriority, SysT
         SysTick__vSetTickCount(u32Tick);
 
         SCB__vRegisterISR(SysTick__vISR,SCB_enVECISR_SYSTICK);
-        SCB_SHPR3->SYSTICK=(uint8_t)enPriority&SCB_SHPR3_SYSTICK_MASK;
-        SysTick_RVR_R=u32Tick-1;
-        SysTick_CVR_R=0;
+        SCB_SHPR3->SYSTICK=(uint32_t)enPriority&SCB_SHPR3_SYSTICK_MASK;
+        SysTick_RVR_R=u32Tick-1u;
+        SysTick_CVR_R=0u;
         SysTick_CSR_R|=SysTick_CSR_R_TICKINT_EN|SysTick_CSR_R_ENABLE_EN;
 
     }
@@ -65,10 +65,10 @@ SysTick_nSTATUS SysTick__enInitTick(uint32_t u32Tick, SCB_nSHPR enPriority, SysT
 SysTick_nSTATUS SysTick__enInitFrequency(float fFrequency, SCB_nSHPR enPriority)
 {
     SysTick_nSTATUS enReturn = SysTick_enOK;
-    uint32_t u32SysTickFrequency=0, u32CountTick=0;
+    uint32_t u32SysTickFrequency=0u, u32CountTick=0;
     float fSysTickFrequency=0.0,fSysTickUs=1000000.0, fCountTick=0.0;
 
-    SysTick_CSR_R=0;
+    SysTick_CSR_R=0u;
     SysTick_CSR->CLKSOURCE=(uint8_t)SysTick_enSYSCLK;
     u32SysTickFrequency=SYSCTL__u32GetClock();
 
@@ -76,16 +76,16 @@ SysTick_nSTATUS SysTick__enInitFrequency(float fFrequency, SCB_nSHPR enPriority)
     fCountTick=fSysTickFrequency/fFrequency;
     u32CountTick=(uint32_t)fCountTick;
 
-    if(u32CountTick>40)
+    if(u32CountTick>40u)
     {
-        if(u32CountTick>(SysTick_VALUEMAX+1))
+        if(u32CountTick>(SysTick_VALUEMAX+1u))
         {
             SysTick_CSR->CLKSOURCE=SysTick_enPIOSC4;
             u32SysTickFrequency=SysTick_PIOSC4;
             fSysTickFrequency=(float)u32SysTickFrequency;
             fCountTick=fSysTickFrequency/fFrequency;
             u32CountTick=(uint32_t)fCountTick;
-            if(u32CountTick>(SysTick_VALUEMAX+1))
+            if(u32CountTick>(SysTick_VALUEMAX+1u))
             {
                 SysTick_vClarAllCounter();
                 enReturn =SysTick_enERROR;
@@ -96,7 +96,7 @@ SysTick_nSTATUS SysTick__enInitFrequency(float fFrequency, SCB_nSHPR enPriority)
             SysTick__vClearFreeCount();
             SysTick__vClearFreeCountOv();
             SysTick__vSetFreqBase(u32SysTickFrequency);
-            SysTick__vSetTimeUsBase(1000000.0/(float)u32SysTickFrequency);
+            SysTick__vSetTimeUsBase(1000000.0f/(float)u32SysTickFrequency);
 
             fSysTickUs/=fFrequency;
 
@@ -105,9 +105,9 @@ SysTick_nSTATUS SysTick__enInitFrequency(float fFrequency, SCB_nSHPR enPriority)
             SysTick__vSetTickCount(u32CountTick);
 
             SCB__vRegisterISR(SysTick__vISR,SCB_enVECISR_SYSTICK);
-            SCB_SHPR3->SYSTICK=(uint8_t)enPriority&SCB_SHPR3_SYSTICK_MASK;
-            SysTick_RVR_R=u32CountTick-1;
-            SysTick_CVR_R=0;
+            SCB_SHPR3->SYSTICK=(uint32_t)enPriority&SCB_SHPR3_SYSTICK_MASK;
+            SysTick_RVR_R=u32CountTick-1u;
+            SysTick_CVR_R=0u;
             SysTick_CSR_R|=SysTick_CSR_R_TICKINT_EN|SysTick_CSR_R_ENABLE_EN;
         }
 
@@ -125,10 +125,10 @@ SysTick_nSTATUS SysTick__enInitFrequency(float fFrequency, SCB_nSHPR enPriority)
 SysTick_nSTATUS SysTick__enInitUs(float fTimeUs, SCB_nSHPR enPriority)
 {
     SysTick_nSTATUS enReturn = SysTick_enOK;
-    uint32_t u32SysTickFrequency=0, u32CountTick=0;
+    uint32_t u32SysTickFrequency=0u, u32CountTick=0;
     float fSysTickFrequency=0.0, fCountTick=0.0;
 
-    SysTick_CSR_R=0;
+    SysTick_CSR_R=0u;
     SysTick_CSR->CLKSOURCE=(uint8_t)SysTick_enSYSCLK;
     u32SysTickFrequency=SYSCTL__u32GetClock();
 
@@ -137,9 +137,9 @@ SysTick_nSTATUS SysTick__enInitUs(float fTimeUs, SCB_nSHPR enPriority)
     fCountTick=fSysTickFrequency*fTimeUs;
     u32CountTick=(uint32_t)fCountTick;
 
-    if(u32CountTick>40)
+    if(u32CountTick>40u)
     {
-        if(u32CountTick>(SysTick_VALUEMAX+1))
+        if(u32CountTick>(SysTick_VALUEMAX+1u))
         {
             SysTick_CSR->CLKSOURCE=SysTick_enPIOSC4;
             u32SysTickFrequency=SysTick_PIOSC4;
@@ -148,7 +148,7 @@ SysTick_nSTATUS SysTick__enInitUs(float fTimeUs, SCB_nSHPR enPriority)
             fCountTick=fSysTickFrequency*fTimeUs;
             u32CountTick=(uint32_t)fCountTick;
 
-            if(u32CountTick>(SysTick_VALUEMAX+1))
+            if(u32CountTick>(SysTick_VALUEMAX+1u))
             {
                 SysTick_vClarAllCounter();
                 enReturn =SysTick_enERROR;
@@ -159,21 +159,21 @@ SysTick_nSTATUS SysTick__enInitUs(float fTimeUs, SCB_nSHPR enPriority)
             SysTick__vClearFreeCount();
             SysTick__vClearFreeCountOv();
             SysTick__vSetFreqBase(u32SysTickFrequency);
-            if(u32SysTickFrequency!=0)
+            if(u32SysTickFrequency!=0u)
             {
-                SysTick__vSetTimeUsBase(1000000.0/(float)u32SysTickFrequency);
+                SysTick__vSetTimeUsBase(1000000.0f/(float)u32SysTickFrequency);
             }
             if(fTimeUs!=0.0)
             {
-                SysTick__vSetTickFreq(1000000.0/fTimeUs);
+                SysTick__vSetTickFreq(1000000.0f/fTimeUs);
             }
             SysTick__vSetTickUs(fTimeUs);
             SysTick__vSetTickCount(u32CountTick);
 
             SCB__vRegisterISR(SysTick__vISR,SCB_enVECISR_SYSTICK);
-            SCB_SHPR3->SYSTICK=(uint8_t)enPriority&SCB_SHPR3_SYSTICK_MASK;
-            SysTick_RVR_R=u32CountTick-1;
-            SysTick_CVR_R=0;
+            SCB_SHPR3->SYSTICK=(uint32_t)enPriority&SCB_SHPR3_SYSTICK_MASK;
+            SysTick_RVR_R=u32CountTick-1u;
+            SysTick_CVR_R=0u;
             SysTick_CSR_R|=SysTick_CSR_R_TICKINT_EN|SysTick_CSR_R_ENABLE_EN;
         }
